+++ prql-compiler/prqlc/src/cli.rs
@@ -6,10 +6,11 @@ use clap::{CommandFactory, Parser, Subcommand, ValueHint};
 use clio::Output;
 use itertools::Itertools;
 use std::io::Write;
+use std::ops::Range;
 use std::path::PathBuf;
 use std::process::exit;
 use std::str::FromStr;
-use std::{fs::File, ops::Range};
+use std::{env, fs::File};
 
 use prql_compiler::semantic::{self, reporting::*};
 use prql_compiler::{ast::pl::Lineage, pl_to_prql};
@@ -27,6 +28,21 @@ pub fn main() -> color_eyre::eyre::Result<()> {
 
     if let Err(error) = cli.command.run() {
         eprintln!("{error}");
+        // Copied from
+        // https://doc.rust-lang.org/src/std/backtrace.rs.html#1-504, since it's private
+        fn backtrace_enabled() -> bool {
+            match env::var("RUST_LIB_BACKTRACE") {
+                Ok(s) => s != "0",
+                Err(_) => match env::var("RUST_BACKTRACE") {
+                    Ok(s) => s != "0",
+                    Err(_) => false,
+                },
+            }
+        }
+        if backtrace_enabled() {
+            eprintln!("{:#}", error.backtrace());
+        }
+
         exit(1)
     }