+++ prql-compiler/src/codegen.rs
@@ -1,4 +1,7 @@
-use crate::{ast::pl, utils::VALID_IDENT};
+use crate::{
+    ast::pl::{self, BinaryExpr},
+    utils::VALID_IDENT,
+};
 
 pub fn write(stmts: &Vec<pl::Stmt>) -> String {
     let mut r = String::new();
@@ -180,7 +183,7 @@ impl WriteSource for pl::ExprKind {
                 }
                 Some(r)
             }
-            Binary { op, left, right } => {
+            Binary(pl::BinaryExpr { op, left, right }) => {
                 let mut r = String::new();
 
                 r += &write_expr(left, self, opt)?;
@@ -198,7 +201,7 @@ impl WriteSource for pl::ExprKind {
             //     pl::UnOp::Not => format!("!{}", expr.write(opt)?),
             //     pl::UnOp::EqSelf => format!("(=={})", expr.write(opt)?),
             // }),
-            Unary { op, expr } => {
+            Unary(pl::UnaryExpr { op, expr }) => {
                 let mut r = String::new();
 
                 // r += &write_expr(left, self, opt)?;
@@ -210,6 +213,12 @@ impl WriteSource for pl::ExprKind {
                 r += &write_expr(expr, self, opt)?;
                 Some(r)
             }
+            // Unary(pl::UnaryExpr { op, expr }) => Some(match op {
+            //     pl::UnOp::Neg => format!("(-{})", expr.write(opt)?),
+            //     pl::UnOp::Add => format!("(+{})", expr.write(opt)?),
+            //     pl::UnOp::Not => format!("!{}", expr.write(opt)?),
+            //     pl::UnOp::EqSelf => format!("(=={})", expr.write(opt)?),
+            // }),
             FuncCall(func_call) => {
                 let mut r = String::new();
                 r += &write_expr(&func_call.name, self, opt)?;
@@ -340,7 +349,7 @@ fn binding_strength(expr: &pl::ExprKind) -> i32 {
         pl::ExprKind::All { .. } => 10,
 
         pl::ExprKind::Range(_) => 6,
-        pl::ExprKind::Binary { op, .. } => match op {
+        pl::ExprKind::Binary(BinaryExpr { op, .. }) => match op {
             pl::BinOp::Mul | pl::BinOp::DivInt | pl::BinOp::DivFloat | pl::BinOp::Mod => 5,
             pl::BinOp::Add | pl::BinOp::Sub => 4,
             pl::BinOp::Eq
@@ -354,7 +363,7 @@ fn binding_strength(expr: &pl::ExprKind) -> i32 {
             pl::BinOp::Or => 1,
             pl::BinOp::Coalesce => 2,
         },
-        pl::ExprKind::Unary { .. } => 6,
+        pl::ExprKind::Unary(..) => 6,
         pl::ExprKind::FuncCall(_) => 7,
         pl::ExprKind::Func(_) => 0,